Moved songs controller to API package
